Align menu icon menu at 250px width logo header is a link to '/' if logged directly show the dashboard otherwise the landing page unselected menu color bold text menu shrink the avatar menu arrow avatar user is a round icon avatar apps are a square icon remove 100% from the authorized app list avatar app is now a link in the apps list remove header border
