+++ src/components/Apps/AuthorizedApps.less
@@ -20,9 +20,10 @@
       a {
         color: #42474d;
       }
-
+      .AppAvatar {
+        height: 60px;
+      }
       .Avatar {
-        border-radius: 50%;
         margin-right: 25px;
       }
     }
