+++ src/components/Auth/RequireLogin.js
@@ -5,7 +5,7 @@ import Loading from '../../widgets/Loading';
 const RequireLogin = ({ auth, location, children }) => {
   if (auth.isLoaded && (!auth.user || (auth.user && !auth.user.name))) {
     const next = location.pathname;
-    const to = next === '/dashboard' ? '/login' : `/login?next=${next}`;
+    const to = next === '/' ? '/login' : `/login?next=${next}`;
     browserHistory.push(to);
   }
   return (auth.isLoaded && auth.user && auth.user.name)
