+++ src/components/LandingPage.less
@@ -22,7 +22,6 @@
   flex-direction: column;
   align-items: center;
   margin-bottom: 180px;
-  overflow: hidden;
 
   @media (max-width: (@screen-sm-max)) {
     min-height: 460px;
